[IMP] - add next_action_type_id to payment return reason form view
Fix payment_return_reason_action_type file name
