+++ account_payment_return/models/__init__.py
@@ -5,3 +5,4 @@ from . import account_invoice
 from . import payment_return_reason
 from . import account_move
 from . import account_journal
+from . import payment_return_reason_action_type
