+++ account_payment_return/models/payment_return_reason.py
@@ -4,23 +4,6 @@
 from odoo import api, fields, models
 
 
-class PaymentReturnReasonActionType(models.Model):
-
-    _name = "payment.return.reason.action.type"
-    _description = "Payment Return Reason Action Types"
-    _order = "sequence, id"
-
-    name = fields.Char(required=True)
-    sequence = fields.Integer("Sequence", default=10)
-    active = fields.Boolean(default=True)
-    company_id = fields.Many2one(
-        "res.company",
-        "Company",
-        default=lambda self: self.env.user.company_id,
-        ondelete="cascade",
-    )
-
-
 class PaymentReturnReason(models.Model):
     _name = "payment.return.reason"
     _description = 'Payment return reason'
